0x1E. C - Search Algorithms

+++ 0x1E-search_algorithms/0-linear.c
@@ -7,20 +7,20 @@
  * @array: input array
  * @size: size of the array
  * @value: value to search in
- * Return: index where value is located, or -1 if value is not present or array is NULL
+ * Return: Always EXIT_SUCCESS
  */
 int linear_search(int *array, size_t size, int value)
 {
-    size_t i;  // Changed from int to size_t for compatibility with size
+	int i;
 
-    if (array == NULL)
-        return (-1);
+	if (array == NULL)
+		return (-1);
 
-    for (i = 0; i < size; i++)
-    {
-        printf("Value checked array[%lu] = [%d]\n", i, array[i]);  // Changed %u to %lu for size_t
-        if (value == array[i])
-            return ((int)i);  // Cast size_t to int
-    }
-    return (-1);
+	for (i = 0; i < (int)size; i++)
+	{
+		printf("Value checked array[%u] = [%d]\n", i, array[i]);
+		if (value == array[i])
+			return (i);
+	}
+	return (-1);
 }
